% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_n.R
\name{ew_n}
\alias{ew_n}
\title{ew_n Function}
\usage{
ew_n(c, utilization)
}
\arguments{
\item{c}{number of servers}

\item{utilization}{utilization of the system}
}
\value{
numeric vector, Returns the expected value of the objects/customers in the SYSTEM.
}
\description{
Expected value (average) of objects in the system
}
\examples{
ew_n(c=3, utilization=0.9)
ew_n(3, 0.9)
}
