% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_nq.R
\name{ew_nq}
\alias{ew_nq}
\title{ew_nq Function}
\usage{
ew_nq(c, utilization)
}
\arguments{
\item{c}{number of servers}

\item{utilization}{utilization of the system}
}
\value{
numeric vector, Returns the expected value of the objects/customers in the QUEUE.
}
\description{
Expected value (average) of objects in the queue
}
\examples{
ew_nq(c= 2, utilization=0.9)
ew_nq(2, 0.9)
}
