% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_w.R
\name{ew_w}
\alias{ew_w}
\title{ew_w Function}
\usage{
ew_w(c, utilization, mu)
}
\arguments{
\item{c}{number of servers}

\item{utilization}{utilization of the system}

\item{mu}{processing/service rate}
}
\value{
numeric vector, Returns the expected time an objects/customers spent in the SYSTEM.
}
\description{
Average time an object spent in the system
}
\examples{
ew_w(c=3, utilization=0.9, mu=3)
ew_w(3, 0.9, 3)
}
