% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_wq.R
\name{ew_wq}
\alias{ew_wq}
\title{ew_wq Function}
\usage{
ew_wq(c, utilization, mu)
}
\arguments{
\item{c}{number of servers}

\item{utilization}{utilization of the system}

\item{mu}{processing/service rate}
}
\value{
numeric vector, Returns the expected time an objects/customers spent in the QUEUE.
}
\description{
Average time an object spent in the queue
}
\examples{
ew_wq(c=3, utilization=0.9, mu=3)
ew_wq(3, 0.9, 3)
}
