% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P0.R
\name{P0}
\alias{P0}
\title{P0 Function}
\usage{
P0(c, u)
}
\arguments{
\item{c}{number of servers}

\item{u}{utilization}
}
\value{
Probability
}
\description{
Returns the probability that there are currently zero objects in the system
}
\examples{
P0(c=4, u=0.75)
}
