% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pn.R
\name{Pn}
\alias{Pn}
\title{Pn Function}
\usage{
Pn(c, u, n)
}
\arguments{
\item{c}{number of servers}

\item{u}{utilization}

\item{n}{number of items}
}
\value{
Probability
}
\description{
Returns the probability of having n items in the system
}
\examples{
Pn(c=5, u=0.85, n=2)
}
