% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PW.R
\name{PW}
\alias{PW}
\title{PW Function}
\usage{
PW(c, lambda, mu, x)
}
\arguments{
\item{c}{number of servers}

\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{x}{time}
}
\value{
Probability
}
\description{
Returns the probability that the cycle time is larger than x
}
\examples{
PW(c=3, lambda=4, mu=5, x=2)
}
