% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWq.R
\name{PWq}
\alias{PWq}
\title{PWq Function}
\usage{
PWq(c, lambda, mu, x)
}
\arguments{
\item{c}{number of servers}

\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{x}{time}
}
\value{
Probability
}
\description{
PWq Function
}
\examples{
PWq(c=3, lambda=4, mu=5, x=1)
}
