% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta.R
\name{zeta}
\alias{zeta}
\title{Zeta Function}
\usage{
zeta(c, utilization)
}
\arguments{
\item{c}{Number of server}

\item{utilization}{Utilization of the system}
}
\value{
Probability
}
\description{
Returns the value of zeta. Zeta is the probability that all servers are busy when you arrive / probability that you have to wait whenn you arrive
}
\examples{
zeta(c=4, utilization=0.8)
zeta(4, 0.8)
}
